feat: reduce error size by boxing some errors (#5577)
### Description
The purpose of this is to reduce memory usage for our agents. When
trying to migrate to rust2024, it warned about certain error types being
unnecessarily large and recommended us to box some error types to reduce
memory footprint.
I set `clippy.toml` during development to:
```toml
large-error-threshold = 64
```
Then ran clippy. Then reduced all the errors that was showing a warning
by boxing them.
VS Code mouse hover was also able to show size of different errors. But
it wasn't able to for some error types :/

I've set it back to 128 in the PR because that's the default value and
setting it to 64 will probably cause a lot of unnecessary warnings.
Unless we feel 128 is too big, we can always set it to something more
appropriate.

### Backward compatibility
Yes

### Testing
None