fix: Move access to database into a blocking task (#5612)
### Description

Move access to database into a blocking task

The access to rocksdb is synchronous and we discovered that it take
about 1 millisecond to make one retrieval of an insertion into a merkle
tree from the database when we build internal merkle tree in relayer
from scratch. As the result, the worker thread which executes the task,
is blocked by access to the database and cannot be used to execute other
scheduled tasks.

This change moves the access to the database into a separate thread in
the blocking pool. It frees up the worker thread to execute other tasks.

### Related issues

- Contributes into https://github.com/hyperlane-xyz/issues/issues/1420

### Backward compatibility

Yes

### Testing

Manual run of relayer locally

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>