fix: Make Sealevel RPC client timeout configurable (#5761)
### Description

Currently we are using Sealevel RPC client with the default timeout
which is 30 seconds. Some of our Sealevel RPC providers take more time
to return a response and the client times out. It is especially relevant
for historical node on Eclipse mainnet where we see timeouts with an RPC
provider.

This change makes the timeout configurable via environment variable
SEALEVEL_RPC_CLIENT_ELEVATED_TIMEOUT_SECONDS so that we can change the
timeout without rebuilding the binary. The configuration impacts all
Sealevel chains.

### Related issues

Timeouts on Eclipse mainnet RPC provider which we use for Scraper and
historical data.

### Backward compatibility

Yes.

If the environment variable is not configured, the current default
timeout will be used.

### Testing

Manual

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>