feat: added cosmos-types package (#5776)
### Description
This PR adds a TypeScript types and encoding/decoding package for the
Hyperlane Cosmos Module. It autogenerates these types and converters
from the proto files from the original [Cosmos SDK Hyperlane
Module](https://github.com/bcp-innovations/hyperlane-cosmos)
implementation. This PR is the result of the discussion from the
previous PR
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5706
implementing the cosmos-sdk. It was decided to move the types in this
cosmos-types package to make it cleaner.

The next steps include creating an SDK which uses those types and
converters to support queries and transaction submission. Furthermore,
using this new SDK as the CosmJs Provider in the Hyperlane SDK and
adding the CosmosHypCollateral and CosmosHypSynthetic token adapter will
be the final step in order to achieve the first milestone of supporting
Cosmos in the Nexus UI.

### How to review
This PR is huge, but only because there is lots of autogenerated code
which can be skipped during the review process. The only code which was
written manually can be found under the `proto` directory and in the
`src/index.ts` file.

### Backward compatibility
Yes

### Testing
None - did manual e2e testing