chore: move rust caching to depot (#5831)
### Description

depot claim:
> we transparently redirect anything bound for GitHub's Actions cache to
our own for extra speed

This PR moves the swatinem/rust-cache backend away from `buildjet` and
back to `github`. As this step takes place on a depot runner, it means
that when trying to cache to `github` it'll actually go to the depot
cache. This means we have far more + quicker available cache for our
agent e2e to work with.

We are still keeping the yarn cache on buildjet for now.

### Drive-by changes

- remove extraneous yarn-cache step from lint-prettier job, as the
yarn-build-with-cache step will automatically pull from the cache if
available
- don't wait for `yarn-install` action before kicking off all the other
jobs. historically we depended on this job to populate the yarn cache if
one did not exist, but ever since
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/4469 we now
check/populate the cache automatically with every
`yarn-build-with-action` step. This will save 1 minute from CI time from
a user-pov, from kicking off all jobs at the start instead of after 1
minute.

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

ci