chore: remove mailboxes from warp configs (#5840)
### Description

This change removes the mailbox addresses from all warp routes that are
written into the registry. Instead of extensive changes to the code
generating the warp configs, we simply remove the mailbox addresses from
the configs before adding them to the registry.

### Related issues

Resolves #5374 

### Backward compatibility

From https://github.com/hyperlane-xyz/hyperlane-registry/pull/679:

> Yes. This change is backward compatible as it only removes redundant
configuration fields that are not used in the deployment process. The
removal of mailbox configurations does not affect the functionality of
the warp routes or their deployment process.

### Testing

Tested by exporting warp configs and comparing with the ones in the
registry branch for
https://github.com/hyperlane-xyz/hyperlane-registry/pull/679