feat: LRU Cache Implementation, ISM and PendingOperation::prepare caching (#4860)
### Description

Implementation of a LRU cache using the moka crate. 
Implementation of MeteredCache exposing cache hit and miss count

Caching ISM functions.
Caching PendingOperation::prepare functions.

Things to note:
1. Storing values which are expensive to clone, if there is a need for
it down the line,
we would need to go by the
[guidelines](https://docs.rs/moka/latest/moka/future/struct.Cache.html#avoiding-to-clone-the-value-at-get):

```
If you want to store values that will be expensive to clone, wrap them by std::sync::Arc before storing in a cache. Arc is a thread-safe reference-counted pointer and its clone() method is cheap.
```

2. We can choose the hash function which will be used for cache key
generation.
The default is `std::hash::RandomState`, but we can choose a different
one by providing a custom hasher.

3. We can change the ttl to be an `Option<ExpirationType>` in all the
set functions,
and resort to the default ttl if the value is `None`. But it may be less
readable in the code
   outside the cache module.

### Drive-by changes

No

### Related issues

Closes #3746

### Backward compatibility

Yes

### Testing

Unit Tests - For Base Cache functionality
Manual Tests - For Metered Cache prometheus metrics

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>
Co-authored-by: Trevor Porter <tkporter4@gmail.com>