feat(submitter): subtask spawner + sealevel e2e (#5868)
### Description

- In the SVM adapter, if `submit` is called on a tx that already has a
`hash`, the submission is skipped. This is because in solana we
shouldn't resubmit unless the hash was dropped, and that would be
detected when status is queried within the inclusion stage, causing the
tx to be dropped, payloads sent back to the building queue, and a new tx
built. This can be optimized but there was no obvious simple way to fix
for now. If we do submit again, a new hash would be generated, creating
a race condition between the old tx and the new tx. In e2e I noticed
that the old tx always lands and the new tx reverts, though the
submitter is unaware of it because its `hash` field gets set to the new
one, so it keeps the tx stuck in the inclusion stage instead of
promoting it to the finality stage.
- Adds PayloadDispatcher `spawn` method, which initializes and connects
all dispatcher components
- Enables the Lander in sealevel e2e for both chains
- Fixes several bugs in the SVM adapter that were causing e2e to fail
- Adds unit testing for the PayloadDispatcherEntrypoint, which found a
bug in the DbLoader
- Adds an integration test that runs the whole Dispatcher (from
entrypoint to finalization), with a mocked adapter
- Adds logs to all stages of the dispatcher, to easily track payloads
and transactions
- Allows configuring `PayloadDispatcherSettings` with a DB instance
rather than a path. This was to fix a bug where the same path cannot be
used by two separate DB instances (the MessageProcessor one and the
Dispatcher one), as an OS lock is acquired on that path by the first DB



### Drive-by changes

- comments out the `gas_expenditure_log_count` e2e invariant because it
is specific to the old submitter
- adds more logs to the e2e invariant checker, to have a better view of
where things fail

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/5824

### Backward compatibility

Yes, but changes to the old sealevel submitter won't be covered in e2e

### Testing

Unit testing and e2e.

Current unit testing coverage report (the columns are Function Coverage
| Line Coverage | Region Coverage)
<img width="1255" alt="Screenshot 2025-04-09 at 17 15 37"
src="https://github.com/user-attachments/assets/258f1730-5275-4ea8-b628-a888b5964d84"
/>