feat: Expose Additional Validator Lifecycle Metrics in Prometheus (#5624)
### Description

This change adds three additional prometheus metrics that are useful for
understanding the health and states of a validator:

- `announced`: Whether the validator has been announced, or is waiting
for announcement in order to sign and submit checkpoints
- `backfill_complete`: Whether the validator has successfully migrated
through the entire checkpoint backlog. Useful to know that the nodes you
are talking to have adequate historical state.
- `reached_initial_consistency`: Whether the validator has hit the chain
head at least once. Useful to know that the validator is in sync (vs
just looking at `contract_sync_height`, where a validator could be
syncing forward but syncing slower than the tip of the chain.

These metrics are a simple 0/1 flag that flip to 1 when the state has
been achieved.

### Testing

Tessellated has these flags deployed in our infrastructure for the last
~3 weeks, and use them in our monitoring to diagnose validator issues