chore(submitter): do not get stuck if a non-retryable error occurs (#5891)
### Description

There was a risk that the submiter gets stuck calling
`retry_until_success` on an action that is guaranteed to be
non-retryable. If that happened, it would prevent liveness - no txs
would be sent to that destination any longer.

This PR adds a new trait, `IsRetryable`, which distinguishes between
retryable and non-retryable `SubmitterError` items. As we gain more
operational experience I'm sure we'll improve the classification, but
for now we should bias towards marking errors as non-retryable (to favor
dropping txs/payloads over halting submissions). We must however be
careful not to drop everything from the submitter as a result of wrongly
classifying an error as non-retryable.

Given the new trait, the PR refactors `retry_until_success` into
`call_until_success_or_nonretryable_error`, which stops retrying if a
nonretryable error is encountered.

### Drive-by changes

- If `BuildingStage::handle_tx_building_result` fails for a transaction
(currently only possible if the inclusion stage channel is full or
dropped), payloads are pushed back to the building queue
- the signature of `ChainTxAdapter::build_transactions` is changed to no
longer return a `Result`, since failures to build the tx will result in
`TxBuildingResult.maybe_tx` to be `None` already

### Related issues

No specific issue, but wanted to make sure we have an easy way to
unblock the submitter if we run into nonretryable errors on RC

### Backward compatibility

Yes

### Testing

Existing sealevel E2E and unit tests