feat: batching optimizations (#5948)
### Description

Relies on https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/5930,
so that one should be merged first

- pre-fetches the RPCs needed to build an EVM multicall batch, to avoid
wastefully making the RPC calls in each subcall preparation. This is
done for the `latest_block` and the `eip1559_fee_estimate`. This is done
just before creating the contract calls for each multicall subcall
- stores the result of the `is_contract` RPC call, whihc was previously
made each time a batch is built
- these are stored in an arc-mutex, inside `EthereumMailbox`
- alone, these improvements lower the time it takes to build the
calldata for a batch from ~2.2s to ~1.1s
- the PR also adds a new configuration field, `bypas_batch_simulation`.
If enabled, it skips the check about whether any subcalls revert inside
the batch and just submits the multicall tx directly. This means the
`confirm` step is relied on to detect that some messages reverted
onchain, and to send those messages back to the prepare queue - this
adds a delay of `CONFIRM_PERIOD` (10 mins) to delivering the reverting
subcalls. This config is disabled by default. When enabled, it saves
another 0.4s, bringing batch calldata building down to 0.7s

### Drive-by changes

- using some of @tkporter's prior testing setup, a unit test meant for
benchmarking is added: `benchmarking_with_real_rpcs`. It logs the RPC
calls being made, and the log timestamps can be used to see where
bottlenecks are. This test is `ignore`d in normal runs
- any logs that were too verbose in `benchmarking_with_real_rpcs` were
simplified. There was a lot of noise, so hopefully this helps lower the
gcp logging bill

### Backward compatibility

Yes, thanks to the new `bypas_batch_simulation` config item

### Testing

Manually using the `benchmarking_with_real_rpcs` test. I ran it with
`bypas_batch_simulation` both enabled and disabled, and made sure the
multicall `eth_call` is skipped when the config is set to `true`

---------

Co-authored-by: Trevor Porter <tkporter4@gmail.com>