feat: Add EvmHypRebaseCollateralAdapter and EvmHypSyntheticRebaseAdaper for monitoring (#5917)
### Description

- Add `EvmHypRebaseCollateralAdapter` and `EvmHypSyntheticRebaseAdapter`
adapters with overriden implementations of `getBridgedSupply`
- `getBridgedSupply` is called by our monitoring code to get the either
the balance of the collateral on the collateral chain or the total
supply on the synthetic side
- When comparing balances, we can either compare the shares on both
sides or the token balance. The shares should remain constant assuming
not transfers happen but the token balances can increase after rebasing.
Either options is valid, we have just chosen to compare shares
- For HypRebaseCollateral type routers, the collateral is actual held in
the 4626 vault not the router contract, and calling `getBalance` on the
vault returns the number of shares
- For the HypSyntheticRebase we are reading the number of shares 


### Backward compatibility


### Testing

Manual