feat(submitter): metrics and e2e invariants (#5904)
### Description

Adds submitter metrics and e2e invariants. The `DispatcherMetrics`
struct only needs the Prometheus `registry` to be instantiated, to avoid
coupling with the rest of the codebase.

I did notice that the adapter is instantiated with the full
`CoreMetrics` unfortunately. If we ever look to fully decouple the
submitter from the rest of the codebase we'll need to fix this, such as
by moving provider metrics logic into their own crate.

Metrics added:
- **task_liveness**: timestamp marking liveness for the stages and db
loaders
- **building_stage_queue_length**
- **inclusion_stage_pool_length**
- **finality_stage_pool_length**
- **dropped_payloads**: includes the reason for dropping
- **dropped_transactions**: includes the reason for dropping
- **finalized_transactions**
- **call_retries**: includes the error message causing the retry.
However, since eyre errors could take on any value, errors are first
simplified as labels that only take on finite values.
- **in_flight_transaction_time**: not maintained yet

### Drive-by changes
- fixes the H256 string formatting bug that made it hard to track the
message id lifecycle
- adds more logs at various steps

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/5819

### Backward compatibility

Yes

### Testing

The new metrics are tested both via unit tests and e2e tests. Failures
are much easier to mock in unit tests.