feat: check new chains using remoteRouters (#5939)
### Description

This PR changes the new chains check implemented in #5914 to allow
non-fully connected warp routes. We now use the `remoteRouters` field in
the deployment configuration. When `remoteRouters` is present, the
checker uses the chains listed there to verify enrollment.

When `remoteRouters` is absent from the deployment config, the checker
assumes the route is fully connected and verifies enrollment for all the
chains in the route.

### Related issues

Resolves #5440 
Fixes ENG-1061

### Backward compatibility

The new behavior of the check is optional, only triggered by
`remoteRouters` field which isn't currently present on any deployment
configs in the registry.

### Testing

Manually running `check-warp-deploy` to verify the check works correctly
with `remoteRouters` on the following routes:

- USDC/arbitrum-base-ethereum-lisk-optimism-polygon-zeronetwork
- USDT/arbitrum-ethereum-mantle-mode-polygon-scroll-zeronetwork