fix(submitter): Cache estimate and use it only on the first submission attempt (#6106)
### Description

Simulation was removed to make Lander as quick and Classical Submitter.
We re-introduce the simulation before submission in the form of
estimation. We shall use the result of the estimation on the first
attempt of submission so that this does not add latency.

### Drive-by changes

Removed the increment of the counter of submission attempts from within
submission logic since currently the value which is produced is not
correct:
- When we retry the transaction submission until success or
non-retryable error, we don't retain the updates from unsuccessful
submissions since we pass a cloned instance of transaction.
- When the submission is successful, we increment the counter in
`process_pending_tx` method.

It means that in all cases the overall result is that the counter is
updated twice.

### Backward compatibility

Yes

### Testing

None

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>