feat: Support commitments in ICA messages (#6051)
### Description

You can now make shielded calls using ICAs. We use a
[commit/reveal](https://en.wikipedia.org/wiki/Commitment_scheme) scheme
where the first message is a commitment, and the second message is a
reveal. The commitment message looks like

```solidity
struct {
  address owner;
  address ism;
  bytes32 salt;
  bytes32 commitment;
}
```

The commitment is `hash(bytes32 salt, Call[] calls)`. 

The reveal message looks like 
```solidity
struct {
  address offChainLookupIsm;
  bytes32 commitment;
}
```

The offChainLookupIsm will contain the urls that a relayer must query to
retrieve the calls/salt given the commitment. The relayer will pass the
retrieved calls/salt to this ism as **metadata**, and the ism will
verify if the salt/calls match the commitment. For more, see #6069 and
the [CCIP spec](https://eips.ethereum.org/EIPS/eip-3668#specification).
### Drive-by changes


### Related issues

https://linear.app/hyperlane-xyz/issue/ENG-1425/ica-call-commitments-smart-contract-changes


### Backward compatibility

No.

### Testing
Unit.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>