feat: agent-specific rpc consensus defaults (#6316)
### Description

We recently discovered that validators use fallback provider as the
default consensus type, which leaves them susceptible to reorgs
regardless of how big their RPC set is.

This PR adds agent-specific rpc consensus defaults: fallback for the
relayer and scraper, quorum for the validator. This default can still be
overriden via the regular config.

I expect this will increase checkpoint-posting latency for validators
that have some bad RPCs in their set.

### Drive-by changes

The `defaultRpcConsensusType` is removed from the agent config schema,
since it applied to all chains and agents, which is much too generic at
this point. It was always set to `fallback` in our infra TS code. It now
still defaults to `fallback` in all cases but when the validator is run.

### Backward compatibility

The agent schema no longer contains `defaultRpcConsensusType`. See
above. This won't break any builds with new agents that use old configs.

### Testing

Manual, by running a local validator and printing the
`origin_chain_conf` to make sure it's using `HttpQuorum`