chore(validator): more verbose merkle tree logic; db integrity check (#6336)
### Description

- makes the log about building the in-memory merkle tree `INFO`, since
some validator operators ignore debug-level logs. We rely on these to
debug reorgs so we must always be able to see them
- adds a bash script that compares merkle leaves from the local DB (read
from the endpoint added in
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/6228) to the
checkpoints posted by a healthy validator to their bucket. The bucket is
currently hardcoded but can be made a parameter later if needed. The
script iterates leaves and stops when it finds the first mismatch in
message ids for a given leaf index

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

mostly: the log wording when ingesting merkle leaves has changed. I'm
not aware of any dependency on that specific log message though.

### Testing

Manual, used these to identify a bad insertion and remedy a reorg