feat: Add reporting of latest checkpoints on reorg at start up (#6361)
### Description

When validator starts up, it creates checkpoint syncer. The syncer
checks if a reorg event was reported by the validator and, if it was,
the syncer constructor will report an error. Validator was intentionaly
crashed (with expect on Result) when it happens.

Now, we capture the error and we report the latest checkpoints as they
are provided by each endpoint separately using reorg reporter. Once
report is logged, we crash the validator.

### Backward compatibility

Yes

### Testing

None

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>