feat(submitter): Naive NonceManager implementation with passing E2E (#6378)
### Description

Improved NonceManager so that it is able to track the status of nonce
and assign proper nonces to transaction. This implementation in the
minimal version of NonceManager which allows E2E to pass with Lander.

- Add database to NonceManager
- Add signer address to NonceManager
- Re-enable Lander Prepare and Confirm tasks in SerialSubmitter
- Remove unnecessary tracking of unfinalised transactions in Finality
stage

More advanced NonceManager will be needed for production use.

### Backward compatibility

Yes

### Testing

E2E Ethereum tests, Unit tests

---------

Co-authored-by: Danil Nemirovsky <4614623+ameten@users.noreply.github.com>