feat(lander): EVM adapter mocking setup (#6450)
### Description

- Changes `EthereumAdapter` so it can more easily be mocked
- Creates a new tests module (`./lander/src/tests/ethereum`) where a
test for the inclusion stage is added. The test successfully gets a
transaction processed, which requires mocking every method in the
`EvmProviderForLander` trait.
- The new test setup serves as scaffolding for reproducing all
submission incidents we've listed in the Lander design doc

Fixes
https://linear.app/hyperlane-xyz/issue/ENG-1710/submitter-evm-integration-testing-setup

### Drive-by changes
- there is some code shift, as tests are moved from
`dispatcher/stages/inclusion_stage.rs` into their own file
- renames two methods each on the TransactionDb and PayloadDb traits,
because there used to be a naming clash which prevents mocking both
traits on the same struct at the same time
- small and inconsequential (you can ignore): slightly changes the
mocking setup for the Lander `Db` - now we use mockall and instead of
storing the state inside the Db struct, we pass it as an arcmutex to the
mocked methods

### Backward compatibility

Yes

### Testing

Unit tests