fix: enforce withdrawal tx sender matches message sender (#6487)
### Description

> In PR #6471, the metadata of the OptimismPortal.WithdrawalTransaction
was modified to include two Hyperlane message IDs (proveMessageId and
finalizeMessageId). However, this patch allows an attacker to create a
separate OP to L1 withdrawal transaction with the same pair of message
IDs in the metadata, and execute OPL2ToL1CcipReadIsm.verify() using that
forged metadata. In other words, it is possible to use the metadata of a
different withdrawal transaction instead of the legitimate one. To
address this issue, OPL2ToL1CcipReadIsm.verify() should verify that the
_tx.sender() from IOptimismPortal.WithdrawalTransaction matches the
sender() of the _message

Fixes this issue by enforcing message sender matches tx sender.

### Backward compatibility

No, requires new deployments

### Testing

Unit Tests